VFB-87: Make Edit Parcel Form load fetch client-side
