+++ src/app/parcels/add/[clientId]/page.tsx
@@ -1,8 +1,11 @@
 import { Metadata } from "next";
 import React from "react";
-import ParcelForm, { ParcelFields } from "@/app/parcels/form/ParcelForm";
-import { Errors, FormErrors } from "@/components/Form/formFunctions";
-import { getCollectionCentresInfo } from "../../form/serverDatabaseFunctions";
+import ParcelForm, {
+    initialParcelFields,
+    initialParcelFormErrors,
+} from "@/app/parcels/form/ParcelForm";
+import supabase from "@/supabaseServer";
+import { getCollectionCentresInfo } from "@/common/fetch";
 
 interface AddParcelParameters {
     params: {
@@ -10,35 +13,16 @@ interface AddParcelParameters {
     };
 }
 
-const initialFields: ParcelFields = {
-    clientId: null,
-    voucherNumber: "",
-    packingDate: null,
-    packingTime: null,
-    shippingMethod: "",
-    collectionDate: null,
-    collectionTime: null,
-    collectionCentre: null,
-};
-
-const initialFormErrors: FormErrors = {
-    voucherNumber: Errors.none,
-    packingDate: Errors.initial,
-    packingTime: Errors.initial,
-    shippingMethod: Errors.initial,
-    collectionDate: Errors.initial,
-    collectionTime: Errors.initial,
-    collectionCentre: Errors.initial,
-};
-
 const AddParcels = async ({ params }: AddParcelParameters): Promise<React.ReactElement> => {
-    const [deliveryPrimaryKey, collectionCentresLabelsAndValues] = await getCollectionCentresInfo();
+    const [deliveryPrimaryKey, collectionCentresLabelsAndValues] = await getCollectionCentresInfo(
+        supabase
+    );
 
     return (
         <main>
             <ParcelForm
-                initialFields={initialFields}
-                initialFormErrors={initialFormErrors}
+                initialFields={initialParcelFields}
+                initialFormErrors={initialParcelFormErrors}
                 clientId={params.clientId}
                 editMode={false}
                 deliveryPrimaryKey={deliveryPrimaryKey}
